Visit doctor bot 1.5 delete bag

+++ src/main/java/com/example/spring_booking_bot/Bot.java
@@ -46,47 +46,51 @@ public class Bot extends TelegramLongPollingBot {
     public void onUpdateReceived(Update update) {
         SendMessage sendMessage = new SendMessage();
         KeyboardRow k = new KeyboardRow();
-        if (userRepo.findUserModelByTgId(update.getMessage().getFrom().getId().toString()) == null) {
-            k.add(new KeyboardButton("Log In"));
-        }
-        k.add(new KeyboardButton("Записаться к врачу"));
-
-        sendMessage.setChatId(update.getMessage().getChatId().toString());
-        sendMessage.setText("выберите действие");
-
-        ReplyKeyboardMarkup replyKeyboardMarkup = new ReplyKeyboardMarkup();
-        replyKeyboardMarkup.setKeyboard(Collections.singletonList(k));
-        sendMessage.setReplyMarkup(replyKeyboardMarkup);
-
-        List<WorkerCommand> list = new ArrayList<>();
-        list.add(new LoginCommand());
-        list.add(new BookCommand());
-        list.add(new TerapevtBookCommand());
-        list.add(new TravmotologBookCommand());
-        list.add(new HirurgBookCommand());
-        list.add(new OkulistBookCommand());
-        list.add(new GinekologBookCommand());
-        list.add(new AllergologBookCommand());
-        list.add(new ChooseTime());
-        list.add(new ViewDoctor());
-        list.add(new AllergologView());
-        list.add(new GinekologView());
-        list.add(new HirurgView());
-        list.add(new OkulistView());
-        list.add(new TerapevtView());
-        list.add(new TravmotologView());
-        list.add(new DeleteCommand());
-        for (WorkerCommand w : list) {
-            sendMessage = w.start(update);
-            if (sendMessage != null) {
-                break;
+        if (update.getMessage().getText().startsWith("/command1") || update.hasMessage()) {
+            if (userRepo.findUserModelByTgId(update.getMessage().getFrom().getId().toString()) == null) {
+                k.add(new KeyboardButton("Log In"));
             }
-        }
+            k.add(new KeyboardButton("Записаться к врачу"));
 
-        try {
-            execute(sendMessage);
-        } catch (TelegramApiException e) {
-            e.printStackTrace();
+            ReplyKeyboardMarkup replyKeyboardMarkup = new ReplyKeyboardMarkup();
+            replyKeyboardMarkup.setKeyboard(Collections.singletonList(k));
+
+            List<WorkerCommand> list = new ArrayList<>();
+            list.add(new LoginCommand());
+            list.add(new BookCommand());
+            list.add(new TerapevtBookCommand());
+            list.add(new TravmotologBookCommand());
+            list.add(new HirurgBookCommand());
+            list.add(new OkulistBookCommand());
+            list.add(new GinekologBookCommand());
+            list.add(new AllergologBookCommand());
+            list.add(new ChooseTime());
+            list.add(new ViewDoctor());
+            list.add(new AllergologView());
+            list.add(new GinekologView());
+            list.add(new HirurgView());
+            list.add(new OkulistView());
+            list.add(new TerapevtView());
+            list.add(new TravmotologView());
+            list.add(new DeleteCommand());
+            for (WorkerCommand w : list) {
+                sendMessage = w.start(update);
+                if (sendMessage != null) {
+                    break;
+                }
+            }
+            if(sendMessage == null) {
+                sendMessage = new SendMessage();
+                sendMessage.setChatId(update.getMessage().getChatId().toString());
+                sendMessage.setReplyMarkup(replyKeyboardMarkup);
+                sendMessage.setText("выберите действие");
+            }
+
+            try {
+                execute(sendMessage);
+            } catch (TelegramApiException e) {
+                e.printStackTrace();
+            }
         }
     }
